backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
python-multipart==0.0.9
pillow==10.4.0
imagehash==4.3.1
transformers==4.44.0
torch==2.4.0
langchain==0.3.0
langchain-community==0.3.0
chromadb==0.5.5
groq==0.11.0
sentence-transformers==3.1.0
google-api-python-client==2.143.0
reportlab==4.2.2
opencv-python-headless==4.10.0.84
numpy==1.26.4
pandas==2.2.2
python-dotenv==1.0.1
aiofiles==24.1.0
httpx==0.27.0